+++ Application.cs
@@ -1,4 +1,5 @@
 using System.Diagnostics;
+using SmartInvestor.Helpers;
 using SmartInvestor.Services.Interfaces;
 
 namespace SmartInvestor;
@@ -10,35 +11,6 @@ public class Application(ICompanyService companyService)
         await companyService.InitializeDatabaseAndDownloadData();
         await companyService.UploadCompanies();
         await companyService.UploadCompanyDtos();
-        StartOptimization();
-    }
-
-    private static void StartOptimization()
-    {
-        var processInfo = new ProcessStartInfo
-        {
-            FileName =  FindPythonInPath(),
-            Arguments = Constants.OptimizationScriptPath,
-            UseShellExecute = true
-        };
-
-        Process.Start(processInfo);
-    }
-
-    private static string FindPythonInPath()
-    {
-        var pathEnv = Environment.GetEnvironmentVariable("PATH") ?? string.Empty;
-        foreach (var dir in pathEnv.Split(';'))
-        {
-            if(!dir.Contains("Python39")) continue;
-            
-            var potentialPythonPath = Path.Combine(dir, "python.exe");
-
-            if (File.Exists(potentialPythonPath))
-            {
-                return potentialPythonPath;
-            }
-        }
-        return string.Empty;
+        OptimizationHelper.StartOptimization();
     }
 }
